scripts/requirements.txt
absl-py==0.11.0
actionlib==1.12.1
angles==1.9.12
astor==0.8.1
attrs==20.3.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
baldor==0.1.2
bleach==3.2.1
bondpy==1.8.5
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.23
catkin-tools==0.6.1
certifi==2020.6.20
chardet==4.0.0
cloudpickle==1.3.0
configparser==4.0.2
contextlib2==0.6.0.post1
controller-manager==0.18.4
controller-manager-msgs==0.18.4
criutils==0.1.3
cv-bridge==1.13.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.7.1
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
distro==1.5.0
docutils==0.16
dynamic-reconfigure==1.6.3
easydict==1.9
empy==3.3.4
entrypoints==0.3
enum-compat==0.0.3
enum34==1.1.10
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.18.2
futures==3.3.0
gast==0.4.0
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.9.0
handeye==0.1.1
html5lib==0.9999999
idna==2.10
image-geometry==1.13.0
importlib-metadata==2.1.1
interactive-markers==1.11.5
ipaddress==1.0.23
ipykernel==4.10.1
ipython==5.10.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
Jinja2==2.11.3
joint-state-publisher==1.12.15
joint-state-publisher-gui==1.12.15
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.5
jupyter-console==5.2.0
jupyter-core==4.6.3
jupyterlab==0.33.12
jupyterlab-launcher==0.11.2
kdl-parser-py==1.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
laser-geometry==1.6.7
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.5
message-filters==1.14.11
mistune==0.8.4
mock==3.0.5
moveit-commander==1.0.8
moveit-core==1.0.8
moveit-ros-planning-interface==1.0.8
moveit-ros-visualization==1.0.8
mpi4py==3.0.3
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.10
numpy==1.16.6
opencv-python==4.0.0.21
osrf-pycommon==0.1.9
packaging==20.9
pandas==0.24.2
pandocfilters==1.4.3
pathlib==1.0.1
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.2
pilz-industrial-motion-testutils==0.6.0
pilz-status-indicator-rqt==0.5.23
prometheus-client==0.9.0
prompt-toolkit==1.0.18
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
pyassimp==4.1.4
pybullet==2.7.9
pydot==1.4.2
Pygments==2.5.2
pyparsing==2.4.7
pyrsistent==0.16.1
python-dateutil==2.8.1
python-qt-binding==0.4.4
pytorch-ignite==0.2.1
pytz==2020.5
PyWavelets==1.0.3
PyYAML==5.4.1
pyzmq==19.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.25.1
resource-retriever==1.12.6
ros-pytest==0.2.0
rosbag==1.14.11
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.11
roslaunch==1.14.11
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.11
rosmake==1.14.9
rosmaster==1.14.11
rosmsg==1.14.11
rosnode==1.14.11
rosparam==1.14.11
rospkg==1.2.10
rospy==1.14.11
rosservice==1.14.11
rostest==1.14.11
rostopic==1.14.11
rosunit==1.14.9
roswtf==1.14.11
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.10
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.7.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.13.18
scandir==1.10.0
scikit-image==0.14.5
scipy==1.2.3
Send2Trash==1.5.0
sensor-msgs==1.12.8
Shapely==1.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
srdfdom==0.5.2
subprocess32==3.5.4
sympy==0.7.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.11
traitlets==4.3.3
trollius==2.2.post1
urdfdom-py==0.4.5
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xacro==1.13.12
zipp==1.2.0
